finishing

+++ app/Http/Controllers/AdminAsistenController.php
@@ -2,12 +2,18 @@
 
 namespace App\Http\Controllers;
 
+use Alert;
 use App\Asisten;
+use App\Instruktur;
+use App\Mail\KirimAkunAsisten;
+use App\Mail\KirimResetPassword;
 use App\User;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Hash;
 use Illuminate\Support\Str;
 use File;
+use Illuminate\Support\Facades\Mail;
+
 class AdminAsistenController extends Controller
 {
     public function index()
@@ -30,13 +36,13 @@ class AdminAsistenController extends Controller
             'npm' => 'required|numeric|unique:App\User,npm',
             'email' => 'required|email|unique:App\User,email',
             'no_telp' => 'numeric|unique:App\User,no_telp',
-            'dir_foto' => 'image'
+            'dir_foto' => 'image|mimes:jpeg,jpg,png'
         ]);
         $pass = Str::random(8);
         $item['password'] = Hash::make($pass);
         $item['role'] = 'asisten';
         $item['active'] = 'y';
-
+        
         $file = $request->file('dir_foto');
         // dd($file);
         if ($file == null) {
@@ -48,6 +54,8 @@ class AdminAsistenController extends Controller
             $item['dir_foto'] = $stored_file->getPathname();
         }
         User::create($item);
+        Mail::to($item['email'])->send(new KirimAkunAsisten($item, $pass));
+        Alert::success('Akun asisten berhasil dibuat', 'Informasi akun telah dikirim ke email '.$item['email']);
         return redirect()->route('admin.asisten.index');
     }
 
@@ -75,13 +83,34 @@ class AdminAsistenController extends Controller
 
     public function delete(Request $request, $id)
     {
-        $item = User::findOrFail($id);
+        $instruktur = Instruktur::where('user_id', $id)->count();
+        if ($instruktur > 0) {
+            Alert::warning('Gagal menghapus akun asisten', 'Asisten tersebut sedang menjadi instruktur praktikum');
+        }else {
+            $item = User::findOrFail($id);
 
-        if ($item->dir_foto != 'assets/admin/img/users/default/user-avatar.jpg') {
-            File::delete($item->dir_foto);
+            if ($item->dir_foto != 'assets/admin/img/users/default/user-avatar.jpg') {
+                File::delete($item->dir_foto);
+            }
+            $item->delete();
+            Alert::success('Akun asisten berhasil dihapus', '');
         }
-        $item->delete();
+        
+        return redirect()->route('admin.asisten.index');
+    }
 
+    public function resetPassword(Request $request, $id)
+    {
+        $data = User::findOrFail($id)->toArray();
+        $pass = Str::random(8);
+        $data['pass'] = $pass;
+        $data['password'] = Hash::make($pass);
+        $user = User::findOrFail($id);
+        $user->update([
+            'password' => $data['password']
+        ]);
+        Mail::to($data['email'])->send(new KirimResetPassword($data));
+        Alert::success('Reset password berhasil dilakukan', 'password telah terkirim ke email '.$data['email']);        
         return redirect()->route('admin.asisten.index');
     }
 }
