+++ app/Http/Controllers/AdminKelasController.php
@@ -2,6 +2,8 @@
 
 namespace App\Http\Controllers;
 
+use Alert;
+use App\Jadwal;
 use App\Kelas;
 use Illuminate\Http\Request;
 
@@ -24,13 +26,17 @@ class AdminKelasController extends Controller
     {
         $item = $request->validate([
             'kelas' => 'required',
-            'jurusan' => 'required',
+            'jurusan' => 'required|not_in:pilih',
             'jumlah_mhs' => ''
         ]);
+        // if ($request->jurusan == "0") {
+        //     return redirect()->route('admin.kelas.create');
+        // }
+
         $item['jumlah_mhs'] = $item['jumlah_mhs'] == null ? 0 : $item['jumlah_mhs'];
 
         Kelas::create($item);
-
+        Alert::success('Jadwal berhasil ditambahkan', '');
         return redirect()->route('admin.kelas.index');
     }
 
@@ -52,15 +58,22 @@ class AdminKelasController extends Controller
         $item['jumlah_mhs'] = $item['jumlah_mhs'] == null ? 0 : $item['jumlah_mhs'];
         $data = Kelas::findOrFail($id);
         $data->update($item);
-
+        Alert::success('Kelas berhasil diubah', '');
         return redirect()->route('admin.kelas.index');
     }
 
     public function delete(Request $request, $id)
     {
-        $item = Kelas::findOrFail($id);
-        $item->delete();
+        $jadwal = Jadwal::where('kelas_id', $id)->count();
 
+        if ($jadwal > 0 ) {
+            Alert::warning('Kelas gagal dihapus', 'Kelas telah terdaftar pada jadwal aktif');
+        }else {
+            $item = Kelas::findOrFail($id);
+            $item->delete();
+            Alert::success('Kelas berhasil dihapus', '');
+        }
+        
         return redirect()->route('admin.kelas.index');
     }
 }
